Add files via upload
This rar file contains all the html, css and javascript associated with the Travello Project. This project was done as an attempt of learning Django. Database used with the project is Postgres. HTML, CSS and Javascript files were taken from https://colorlib.com/wp/template/travello/ . Attempt was successful attempt of making the website dynamic, we can add files from database and also user login and registration can be done with updating the database as well.